Some DB changes and other cleanup

+++ bin/make_db.py
@@ -1,24 +1,37 @@
-import sys, os
-
+#!/usr/bin/env python
+__author__ = 'Jon Stratton'
+import sys, os, getopt
 run_dir = os.path.dirname(os.path.realpath(__file__))
 lib_loc = os.path.realpath('%s/../lib/' % run_dir)
 sys.path.insert(1, lib_loc )
 import mesh_front_util as mfu
+import mesh_front_db as mfdb
 import sqlite3
 
+# Defaults
 password = 'changeme'
-password_hash = mfu.hash_password(password).hexdigest()
-#print(password_hash)
+refresh  = 0
 
+# Command line params
+myopts, args = getopt.getopt(sys.argv[1:],'p:r')
+for o, a in myopts:
+    if (o == '-p'):
+        password = a
+    elif (o == '-r'):
+        refresh = 1
+
+
+password_hash = mfu.hash_password(password).hexdigest()
 conn = sqlite3.connect('db.sqlite3')
 c = conn.cursor()
 
 # Drop tables if refresh
-c.execute('DROP TABLE user_settings;')
-c.execute('DROP TABLE interface_settings;')
-c.execute('DROP TABLE mesh_settings;')
-c.execute('DROP TABLE server_settings;')
-c.execute('DROP TABLE service_settings;')
+if (refresh):
+    c.execute('DROP TABLE user_settings;')
+    c.execute('DROP TABLE interface_settings;')
+    c.execute('DROP TABLE mesh_settings;')
+    c.execute('DROP TABLE server_settings;')
+    #c.execute('DROP TABLE service_settings;')
 
 # Admin passwd table
 c.execute('CREATE TABLE user_settings (username text, password_hash text);')
@@ -34,7 +47,12 @@ c.execute('CREATE TABLE mesh_settings (key text, value text);')
 
 # Generic server settings: hostname, wifi_channel, wifi_ssid, dns1, dns2, share_iternet
 c.execute('CREATE TABLE server_settings (key text, value text);')
-c.execute('INSERT INTO server_settings VALUES (\'hostname\', ?);', (mfu.get_hostname(),) )
+mfdb.set_setting('hostname', mfu.get_hostname())
+mfdb.set_setting('listen_port', '8080')
+mfdb.set_setting('listen_ip', '0.0.0.0')
+#c.execute('INSERT INTO server_settings VALUES (\'hostname\', ?);', (mfu.get_hostname(),) )
+#c.execute('INSERT INTO server_settings VALUES (\'listen_port\', ?);', (8080,) )
+#c.execute('INSERT INTO server_settings VALUES (\'listen_ip\', ?);', ('0.0.0.0',) )
 
 #c.execute('SELECT * FROM users')
 #print(c.fetchone())
